+++ opendrive/src/lane/offset.rs
@@ -5,17 +5,17 @@ use serde::{Deserialize, Serialize};
 pub struct Offset {
     /// Polynom parameter a, offset at @s (ds=0)
     #[serde(rename = "@a")]
-    pub a: f32,
+    pub a: f64,
     /// Polynom parameter b
     #[serde(rename = "@b")]
-    pub b: f32,
+    pub b: f64,
     /// Polynom parameter c
     #[serde(rename = "@c")]
-    pub c: f32,
+    pub c: f64,
     /// Polynom parameter d
     #[serde(rename = "@d")]
-    pub d: f32,
+    pub d: f64,
     /// s-coordinate of start position
     #[serde(rename = "@s")]
-    pub s: f32,
+    pub s: f64,
 }
