+++ opendrive/src/util.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Deserializer};
 
-pub fn deserialize_scientific<'de, D>(deserializer: D) -> Result<f32, D::Error>
+pub fn deserialize_scientific<'de, D>(deserializer: D) -> Result<f64, D::Error>
 where
     D: Deserializer<'de>,
 {
@@ -8,7 +8,7 @@ where
 
     String::deserialize(deserializer).and_then(|string| {
         let float = string
-            .parse::<f32>()
+            .parse::<f64>()
             .map_err(|err| Error::custom(err.to_string()))?;
 
         Ok(float)
